Introduce a setup command
